requirements.txt
fastapi~=0.115.2
uvicorn~=0.32.0
motor~=3.6.0
python-dotenv~=1.0.1
requestez~=0.1.5
beautifulsoup4~=4.12.3
requests~=2.32.3
starlette~=0.40.0
pycryptodome~=3.12.0
